experience modified

+++ server_api/src/modules/applicants/positions/position.service.ts
@@ -1,5 +1,6 @@
 import { AppError } from "../../../common/errors/app.error";
 import { ERROR_CODES } from "../../../common/errors/error.codes";
+import { ExperienceRepository, IExperienceRepository } from "../experiences/experience.repository";
 import {
     CreatePositionDTO,
     GetPositionsDTO,
@@ -11,7 +12,8 @@ import { IPositionRepository, PositionRepository } from "./position.repository";
 export class PositionService {
 
     constructor(
-        private readonly positionRepo: IPositionRepository = new PositionRepository()
+        private readonly positionRepo: IPositionRepository = new PositionRepository(),
+        private readonly experienceRepo: IExperienceRepository = new ExperienceRepository()
     ) { }
 
     /* =========================
@@ -61,6 +63,10 @@ export class PositionService {
     async delete(id: string) {
         const hasRanks = await this.positionRepo.exists({ parent: id });
         if (hasRanks) throw new AppError(ERROR_CODES.RANK_ALREADY_EXISTS);
+        const isUsedInExperience = await this.experienceRepo.exists({ rank: id });
+        if (isUsedInExperience) {
+            throw new AppError(ERROR_CODES.RANK_IN_USE);
+        }
         const deleted = await this.positionRepo.delete(id);
         if (!deleted) throw new AppError(ERROR_CODES.POSITION_NOT_FOUND);
         return deleted;
